refactor(#1245): split calendar-events.ts into calendar-view + events
calendar-events.ts hosted both the timeline/grid view-and-range plumbing and
the booking/block CalendarItem union transforms (co-evolved via #1204 + #1101).
Split by concern:

- calendar-view.ts (new) — CalendarView, view gating (operatorViews/
  defaultCalendarView), calendarRange/shiftCalendarDate, and the URL <-> state
  params (normalizeViewParam/parseCalendarView/format+parseCalendarDate).
- calendar-events.ts (trimmed) — the CalendarItem union types +
  toCalendarEvents/blocksToCalendarEvents/fleetToResources/calendarItemClassName.

The two modules have no cross-import. Consumers (BookingsCalendar, FleetTimeline,
CalendarToolbar, the bookings route) re-pointed per symbol; the calendar-events
test file split alongside into calendar-view.test.ts + calendar-events.test.ts.

Folds in the #1245 LOW: useCalendarFilters.filterEvents widened its generic to
`<T extends { resourceId: string; status?: OperatorBookingStatus }>` so the
structural `as`-cast is gone (a block omits status, still assignable).

Behavior-preserving: web typecheck clean, full web suite 1853/1853 green.

Refs #1099 #1101
Closes #1245
